Update README to current project status
This got stalled a year and a half ago, when I discovered that the
pseudo-code for splitting and recombination wasn't much more than a
guess. I sat on it for a bit to see if I could think of an approach
that might work, then other matters ate my time.

Now I'm working on a port of xu88/Green to something more resistant to
memory issues than C (D language). Roger has told me that the
recombine part of Green worked (and it took him awhile to get it
right) so when I get to that part we'll see if I can salvage this
project. In the meantime if someone else wants to try, go ahead. Just
let me (jdougan@acm.org) know. --John Dougan